requirements.txt

Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
Pillow==7.2.0
requests==2.24.0

